+++ internal/command/cliconfig/cliconfig.go
@@ -9,11 +9,14 @@
 package cliconfig
 
 import (
+	"errors"
 	"fmt"
+	"io/fs"
 	"io/ioutil"
 	"log"
 	"os"
 	"path/filepath"
+	"strings"
 
 	"github.com/hashicorp/hcl"
 
@@ -22,6 +25,7 @@ import (
 )
 
 const pluginCacheDirEnvVar = "TF_PLUGIN_CACHE_DIR"
+const pluginCacheMayBreakLockFileEnvVar = "TF_PLUGIN_CACHE_MAY_BREAK_DEPENDENCY_LOCK_FILE"
 
 // Config is the structure of the configuration for the Terraform CLI.
 //
@@ -100,12 +104,14 @@ func LoadConfig() (*Config, tfdiags.Diagnostics) {
 	configVal := BuiltinConfig // copy
 	config := &configVal
 
-	if mainFilename, err := cliConfigFile(); err == nil {
+	if mainFilename, mainFileDiags := cliConfigFile(); len(mainFileDiags) == 0 {
 		if _, err := os.Stat(mainFilename); err == nil {
 			mainConfig, mainDiags := loadConfigFile(mainFilename)
 			diags = diags.Append(mainDiags)
 			config = config.Merge(mainConfig)
 		}
+	} else {
+		diags = diags.Append(mainFileDiags)
 	}
 
 	// Unless the user has specifically overridden the configuration file
@@ -220,9 +226,14 @@ func loadConfigDir(path string) (*Config, tfdiags.Diagnostics) {
 // Any values specified in this config should override those set in the
 // configuration file.
 func EnvConfig() *Config {
+	env := makeEnvMap(os.Environ())
+	return envConfig(env)
+}
+
+func envConfig(env map[string]string) *Config {
 	config := &Config{}
 
-	if envPluginCacheDir := os.Getenv(pluginCacheDirEnvVar); envPluginCacheDir != "" {
+	if envPluginCacheDir := env[pluginCacheDirEnvVar]; envPluginCacheDir != "" {
 		// No Expandenv here, because expanding environment variables inside
 		// an environment variable would be strange and seems unnecessary.
 		// (User can expand variables into the value while setting it using
@@ -230,7 +241,32 @@ func EnvConfig() *Config {
 		config.PluginCacheDir = envPluginCacheDir
 	}
 
+	if envMayBreak := env[pluginCacheMayBreakLockFileEnvVar]; envMayBreak != "" && envMayBreak != "0" {
+		// This is an environment variable analog to the
+		// plugin_cache_may_break_dependency_lock_file setting. If either this
+		// or the config file setting are enabled then it's enabled; there is
+		// no way to override back to false if either location sets this to
+		// true.
+		config.PluginCacheMayBreakDependencyLockFile = true
+	}
+
 	return config
+}
+
+func makeEnvMap(environ []string) map[string]string {
+	if len(environ) == 0 {
+		return nil
+	}
+
+	ret := make(map[string]string, len(environ))
+	for _, entry := range environ {
+		eq := strings.IndexByte(entry, '=')
+		if eq == -1 {
+			continue
+		}
+		ret[entry[:eq]] = entry[eq+1:]
+	}
+	return ret
 }
 
 // Validate checks for errors in the configuration that cannot be detected
@@ -328,6 +364,12 @@ func (c *Config) Merge(c2 *Config) *Config {
 		result.PluginCacheDir = c2.PluginCacheDir
 	}
 
+	if c.PluginCacheMayBreakDependencyLockFile || c2.PluginCacheMayBreakDependencyLockFile {
+		// This setting saturates to "on"; once either configuration sets it,
+		// there is no way to override it back to off again.
+		result.PluginCacheMayBreakDependencyLockFile = true
+	}
+
 	if (len(c.Hosts) + len(c2.Hosts)) > 0 {
 		result.Hosts = make(map[string]*ConfigHost)
 		for name, host := range c.Hosts {
@@ -369,7 +411,8 @@ func (c *Config) Merge(c2 *Config) *Config {
 	return &result
 }
 
-func cliConfigFile() (string, error) {
+func cliConfigFile() (string, tfdiags.Diagnostics) {
+	var diags tfdiags.Diagnostics
 	mustExist := true
 
 	configFilePath := cliConfigFileOverride()
@@ -389,15 +432,19 @@ func cliConfigFile() (string, error) {
 	f, err := os.Open(configFilePath)
 	if err == nil {
 		f.Close()
-		return configFilePath, nil
+		return configFilePath, diags
 	}
 
-	if mustExist || !os.IsNotExist(err) {
-		return "", err
+	if mustExist || !errors.Is(err, fs.ErrNotExist) {
+		diags = append(diags, tfdiags.Sourceless(
+			tfdiags.Warning,
+			"Unable to open CLI configuration file",
+			fmt.Sprintf("The CLI configuration file at %q does not exist.", configFilePath),
+		))
 	}
 
 	log.Println("[DEBUG] File doesn't exist, but doesn't need to. Ignoring.")
-	return "", nil
+	return "", diags
 }
 
 func cliConfigFileOverride() string {