+++ CHANGELOG.md
@@ -13,6 +13,10 @@ ENHANCEMENTS:
 * Terraform CLI's local operations mode will now attempt to persist state snapshots to the state storage backend periodically during the apply step, thereby reducing the window for lost data if the Terraform process is aborted unexpectedly. [GH-32680]
 * If Terraform CLI recieves SIGINT (or its equivalent on non-Unix platforms) during the apply step then it will immediately try to persist the latest state snapshot to the state storage backend, with the assumption that a graceful shutdown request often typically followed by a hard abort some time later if the graceful shutdown doesn't complete fast enough. [GH-32680]
 
+BUG FIXES:
+
+* `terraform init`: Fixed crash with invalid blank module name. [GH-32781]
+
 ## Previous Releases
 
 For information on prior major and minor releases, see their changelogs: